move the reloadFunction
