+++ packages/packages/core/editor-global-classes/src/components/class-manager/delete-class.ts
@@ -1,5 +1,3 @@
-import { getCurrentDocument, getV1DocumentsManager } from '@elementor/editor-documents';
-import { __privateRunCommand as runCommand } from '@elementor/editor-v1-adapters';
 import { __dispatch as dispatch } from '@elementor/store';
 
 import { slice } from '../../store';
@@ -23,18 +21,3 @@ export const onDelete = async () => {
 };
 
 export const hasDeletedItems = () => isDeleted;
-
-// When deleting a class, we remove it from all the documents that have it applied.
-// In order to reflect the changes in the active document, we need to reload it.
-export const reloadDocument = () => {
-	const currentDocument = getCurrentDocument();
-	const documentsManager = getV1DocumentsManager();
-
-	documentsManager.invalidateCache();
-
-	return runCommand( 'editor/documents/switch', {
-		id: currentDocument?.id,
-		shouldScroll: false,
-		shouldNavigateToDefaultRoute: false,
-	} );
-};
